// fe/src/lib/api/generated/model/csvImportResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { CsvImportRowResponse } from './csvImportRowResponse';

export interface CsvImportResponse {
  committed: boolean;
  conflict_count: number;
  create_count: number;
  invalid_count: number;
  rows: CsvImportRowResponse[];
  skip_count: number;
  total_rows: number;
  valid: boolean;
}